get database initial settings in layout.tsx only on init app load
